+++ new_project/examples/separate_apps/src/maker.ts
@@ -1,7 +1,29 @@
 import { formatEther } from "ethers/utils";
+import moment from "moment";
 import readLineSync from "readline-sync";
 import { toBitcoin } from "satoshi-bitcoin-ts";
 import { checkEnvFile, startClient } from "./lib";
+import { NegotiationProtocolHandler, Order } from "./negotiation";
+
+function createOrder(): Order {
+    return {
+        id: "123",
+        key: "ETH-BTC",
+        valid_until: moment().unix() + 300,
+        ask: {
+            amount: "9000000000000000000",
+            asset: "ether",
+            ledger: "ethereum",
+            network: "regtest",
+        },
+        bid: {
+            amount: "100000000",
+            asset: "bitcoin",
+            ledger: "bitcoin",
+            network: "regtest",
+        },
+    };
+}
 
 (async function main() {
     checkEnvFile(process.env.DOTENV_CONFIG_PATH!);
@@ -24,13 +46,30 @@ import { checkEnvFile, startClient } from "./lib";
     console.log("[Maker] peer id:", peerId);
     console.log("[Maker] address hint:", addressHint);
 
-    // Note that we assume the maker published an offer
-    // somehow somewhere and that someone (a taker)
-    // will take such offer by send a SWAP request to the
-    // maker's COMIT node. Publishing and Finding offers
-    // is not part of this example.
+    // start negotiation protocol handler so that a taker can take the order and receives the latest rate
 
-    console.log("Waiting for someone to take my offer...");
+    const negotiationProtocolHandler = new NegotiationProtocolHandler(
+        {
+            connection_info: {
+                peer_id: peerId,
+                address_hint: addressHint,
+            },
+            expiries: {
+                ask_expiry: moment().unix() + 7200,
+                bid_expiry: moment().unix() + 3600,
+            },
+            role: "alice",
+            swap_id: "SOME_RANDOM_ID",
+        },
+        2318
+    ); // CoBloX Founding Date 🚀
+
+    negotiationProtocolHandler.start();
+    const order = createOrder();
+    negotiationProtocolHandler.addOrder(order);
+
+    const invitationDetails = `http://localhost:2318/orders/ETH-BTC`;
+    console.log(`Waiting for someone taking my order at: ${invitationDetails}`);
 
     let swapHandle;
     while (!swapHandle) {
@@ -45,10 +84,19 @@ import { checkEnvFile, startClient } from "./lib";
     }
 
     const actionConfig = { timeout: 100000, tryInterval: 1000 };
-    await swapHandle.accept(actionConfig);
 
     const swap = await swapHandle.getEntity();
     const swapParams = swap.properties!.parameters;
+
+    // only accept a request if it fits to the created order above
+    if (isValid(swapParams, order)) {
+        console.log("Requested order is invalid");
+        await swapHandle.decline(actionConfig);
+        process.exit();
+    }
+    console.log("Requested order is still valid");
+    await swapHandle.accept(actionConfig);
+
     console.log(
         "Swap started! Swapping %d Ether for %d %s",
         formatEther(swapParams.alpha_asset.quantity),
@@ -80,3 +128,11 @@ import { checkEnvFile, startClient } from "./lib";
     );
     process.exit();
 })();
+
+function isValid(swapParams: any, order: Order) {
+    return (
+        swapParams.alpha_asset.name !== order.ask.asset ||
+        swapParams.beta_asset.name !== order.bid.asset ||
+        order.valid_until < moment().unix()
+    );
+}