+++ new_project/examples/separate_apps/src/taker.ts
@@ -1,9 +1,13 @@
 import { SwapRequest } from "comit-sdk";
 import { formatEther } from "ethers/utils";
-import moment from "moment";
 import readLineSync from "readline-sync";
-import { toBitcoin, toSatoshi } from "satoshi-bitcoin-ts";
+import { toBitcoin } from "satoshi-bitcoin-ts";
 import { Actor, checkEnvFile, startClient } from "./lib";
+import {
+    ExecutionParams,
+    NegotiationProtocolClient,
+    Order,
+} from "./negotiation";
 
 (async function main() {
     checkEnvFile(process.env.DOTENV_CONFIG_PATH!);
@@ -18,20 +22,38 @@ import { Actor, checkEnvFile, startClient } from "./lib";
         ).toFixed(2)
     );
 
-    const peerId = readLineSync.question("What is the Maker's peer id?");
-    const addressHint = readLineSync.question(
-        "What is the Maker's address hint?"
+    readLineSync.question("0. Ready?");
+
+    // take an order from a maker
+    const negotiationProtocolClient = new NegotiationProtocolClient();
+    const {
+        order,
+        execution_params,
+    } = await negotiationProtocolClient.startNegotiation(
+        "http://localhost:2318/orders",
+        "ETH-BTC"
     );
 
-    const swapMessage = createSwap(taker, peerId, addressHint);
+    const ether = formatEther(order.ask.amount);
+    const bitcoin = toBitcoin(order.bid.amount);
+    console.log(
+        `Received latest order details: %s:%s for a rate of %d:%d`,
+        order.ask.asset,
+        order.bid.asset,
+        ether,
+        bitcoin
+    );
+
+    const swapMessage = createSwap(taker, order, execution_params);
 
     const swapHandle = await taker.comitClient.sendSwap(swapMessage);
 
     const actionConfig = { timeout: 100000, tryInterval: 1000 };
 
     console.log(
-        "Swap started! Swapping %d Ether for %d %s",
+        "Swap started! Swapping %d %s for %d %s",
         formatEther(swapMessage.alpha_asset.quantity),
+        swapMessage.alpha_asset.name,
         toBitcoin(swapMessage.beta_asset.quantity),
         swapMessage.beta_asset.name
     );
@@ -63,34 +85,34 @@ import { Actor, checkEnvFile, startClient } from "./lib";
 
 function createSwap(
     actor: Actor,
-    peerId: string,
-    addressHint: string
+    order: Order,
+    executionParams: ExecutionParams
 ): SwapRequest {
     const refundAddress = actor.ethereumWallet.getAccount();
 
     return {
         alpha_ledger: {
-            name: "ethereum",
-            network: "regtest",
+            name: order.ask.ledger,
+            network: order.ask.network,
         },
         beta_ledger: {
-            name: "bitcoin",
-            network: "regtest",
+            name: order.bid.ledger,
+            network: order.bid.network,
         },
         alpha_asset: {
-            name: "ether",
-            quantity: "9000000000000000000",
+            name: order.ask.asset,
+            quantity: order.ask.amount,
         },
         beta_asset: {
-            name: "bitcoin",
-            quantity: toSatoshi(1).toString(),
+            name: order.bid.asset,
+            quantity: order.bid.amount,
         },
         alpha_ledger_refund_identity: refundAddress,
-        alpha_expiry: moment().unix() + 7200,
-        beta_expiry: moment().unix() + 3600,
+        alpha_expiry: executionParams.expiries.ask_expiry,
+        beta_expiry: executionParams.expiries.bid_expiry,
         peer: {
-            peer_id: peerId,
-            address_hint: addressHint,
+            peer_id: executionParams.connection_info.peer_id,
+            address_hint: executionParams.connection_info.address_hint,
         },
     };
 }